VOL-2800 fix for missing DHCP flows on NNI port after OLT delete and recreation.
Sometimes the NNI port can be in disabled state when VOLTHA connects to ONOS.
Because the port exists but is disabled at this point, NNI flows will not get programmed.
Also the port will not be `added` later as it is already there.
So again NNI flows will not be programmed as there will be no PORT_ADDED message.
As a result, we must handle the PORT_UPDATED case for NNI ports as well.

Change-Id: Ie9a7fcbe91a8181c54fcfa3aa5ae5c26bc64bb0e