[VOL-3086][VOLT-3087] Removing flows from all ONOS instances, withdrawing meters
Every instance needs to explicilty purge the local DeviceFlowTable cache because the
FlowRuleManager does not comunicate with the cluster the request to remove all flows
when a device is disconnected.

With this patch also Meters get into PENDING_REMOVE state and get removed upon
VOLTHA/OLT re-connection. A better implementation is to have a new API in ONOS
to forcefully remove the meters.

Change-Id: Ic7afa06e5d7e71017e9b8b795d341058ecbfd054