[VOL-4513] Always push events on the queue, even if they are duplicates
In the case of an Add -> Delete -> Add that are received very quickly,
we might try to add the second Add to the queue before the first one is removed.

If that's the case the second Add won't be processed and we'll be left in an inconsistent state.
The event handling is idempotent anyway so there's no issue in duplicating them.

Change-Id: Iebc28a147eff9a061148bc72a9878b52a62e113d