Stopping considering default EAPOL with last status 'ERROR' as present
We noticed that during some jobs execution (e.g: onu mib template unknown me) the default EAPOL flow would not be reinstalled
after running into error during a previous uninstallation. The reason was because of the verification function 'hasDefaultEapol'.
It was considering the last status ERROR when evaluating if the default EAPOL flow was present. There was a comment justifying that rule,
however the premise was proven to be false. This patchset was made on that context.

Change-Id: Ifc3d802a3bc8e45ff75565647faa11d3ed212f8f
Signed-off-by: Gustavo Silva <gsilva@furukawalatam.com>